requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --frozen --output-file=requirements.txt --no-hashes
-e .
absl-py==2.1.0
accelerate==0.31.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==5.0.1 ; python_full_version < '3.11'
atomicwrites==1.4.1 ; sys_platform == 'win32'
attrs==23.2.0
audioop-lts==0.2.1 ; python_full_version >= '3.13'
audioread==3.0.1
bitsandbytes==0.43.1
cachetools==5.5.0
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6 ; sys_platform == 'win32' or platform_system == 'Windows'
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
decorator==5.1.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
docstring-parser==0.16
einops==0.8.0
exceptiongroup==1.2.1 ; python_full_version < '3.11'
fastapi==0.115.11
ffmpy==0.3.2
filelock==3.15.4
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.2/flash_attn-2.5.2+cu118torch2.2cxx11abiFALSE-cp310-cp310-linux_x86_64.whl
fonttools==4.53.0
frozenlist==1.5.0
fsspec==2024.5.0
future==1.0.0
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.82.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gradio==5.20.0
gradio-client==1.7.2
groovy==0.1.2
grpc-google-iam-v1==0.14.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.29.1
identify==2.6.3
idna==3.7
iniconfig==2.0.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
levenshtein==0.25.1
librosa==0.9.2
llvmlite==0.43.0
markdown==3.7
markdown-it-py==2.2.0 ; sys_platform != 'emscripten'
markupsafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2 ; sys_platform != 'emscripten'
memoization==0.4.0
mir-eval==0.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-cupti-cu11==11.8.87 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-nvrtc-cu11==11.8.89 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-runtime-cu11==11.8.89 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cudnn-cu11==8.7.0.84 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cufft-cu11==10.9.0.58 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-curand-cu11==10.3.0.86 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cusolver-cu11==11.4.1.48 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cusparse-cu11==11.7.5.86 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-nccl-cu11==2.19.3 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-nvtx-cu11==11.8.86 ; platform_machine == 'x86_64' and platform_system == 'Linux'
orjson==3.10.5
packaging==24.1
pandas==1.4.3
peft==0.11.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
plumbum==1.7.2
pooch==1.8.2
pre-commit==4.0.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.27.2
psutil==6.0.0
py==1.11.0
pyarrow==19.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.7.4
pydantic-core==2.18.4
pydantic-settings==2.7.1
pydub==0.25.1
pygments==2.18.0 ; sys_platform != 'emscripten'
pyparsing==3.1.2
pytest==7.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
pywin32==308 ; platform_python_implementation != 'PyPy' and platform_system == 'Windows'
pyyaml==6.0
rapidfuzz==3.9.4
regex==2024.5.15
requests==2.32.3
resampy==0.3.1
rich==13.7.1 ; sys_platform != 'emscripten'
rsa==4.9
ruff==0.9.9
safehttpx==0.1.6
safetensors==0.4.3
scikit-learn==1.1.1
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==75.6.0
shapely==2.0.7
shellingham==1.5.4 ; sys_platform != 'emscripten'
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
starlette==0.46.0
sympy==1.12.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.2.2 ; platform_system == 'Darwin'
torch==2.2.2 ; platform_system != 'Darwin' and platform_system != 'Linux'
torch==2.2.2+cu118 ; platform_system == 'Linux'
torchaudio==2.2.2 ; platform_system == 'Darwin'
torchaudio==2.2.2 ; platform_system != 'Darwin' and platform_system != 'Linux'
torchaudio==2.2.2+cu118 ; platform_system == 'Linux'
torchvision==0.17.2 ; platform_system == 'Darwin'
torchvision==0.17.2 ; platform_system != 'Darwin' and platform_system != 'Linux'
torchvision==0.17.2+cu118 ; platform_system == 'Linux'
tqdm==4.66.4
transformers==4.44.2
triton==2.2.0 ; python_full_version < '3.12' and platform_machine == 'x86_64' and platform_system == 'Linux'
typer==0.12.3 ; sys_platform != 'emscripten'
typing-extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1 ; sys_platform != 'emscripten'
virtualenv==20.28.0
wandb==0.17.3
websockets==12.0
werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
